Integrate the new data format with the library
