+++ src/lib.rs
@@ -54,6 +54,12 @@ pub mod algorithms;
 pub mod routines {
     /// Handles datafile operations
     pub mod datafile;
+    /// New data format
+    pub mod data {
+        pub mod parse_pmetrics;
+        pub mod structures;
+        pub mod traits;
+    }
     /// Routines for initializing the grid
     pub mod initialization;
     pub mod optimization {
